+++ src/main/java/net/mcreator/element/converter/ConverterRegistry.java
@@ -31,12 +31,7 @@ import net.mcreator.element.converter.v2020_5.BiomeDefaultFeaturesConverter;
 import net.mcreator.element.converter.v2020_5.GUICoordinateConverter;
 import net.mcreator.element.converter.v2020_5.OverlayCoordinateConverter;
 import net.mcreator.element.converter.v2020_5.ProcedureSpawnGemPickupDelayFixer;
-import net.mcreator.element.converter.v2020_6.BlockLuminanceFixer;
-import net.mcreator.element.converter.v2020_6.DimensionLuminanceFixer;
-import net.mcreator.element.converter.v2020_6.PlantLuminanceFixer;
-import net.mcreator.element.converter.v2021_1.BiomeFrozenTopLayerConverter;
-import net.mcreator.element.converter.v2021_1.BlockBoundingBoxFixer;
-import net.mcreator.element.converter.v2021_1.GameruleDisplayNameFixer;
+import net.mcreator.element.converter.v2021_1.*;
 import net.mcreator.element.converter.v2021_2.*;
 import net.mcreator.element.converter.v2021_3.LegacyProcedureBlockRemover;
 import net.mcreator.element.converter.v2022_1.FoodToItemConverter;
