+++ src/main/java/net/mcreator/element/converter/v2021_1/BlockLuminanceFixer.java
@@ -17,32 +17,32 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.mcreator.element.converter.v2020_6;
+package net.mcreator.element.converter.v2021_1;
 
 import com.google.gson.JsonElement;
 import net.mcreator.element.GeneratableElement;
 import net.mcreator.element.converter.IConverter;
-import net.mcreator.element.types.Plant;
+import net.mcreator.element.types.Block;
 import net.mcreator.workspace.Workspace;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-public class PlantLuminanceFixer implements IConverter {
-	private static final Logger LOG = LogManager.getLogger(PlantLuminanceFixer.class);
+public class BlockLuminanceFixer implements IConverter {
+	private static final Logger LOG = LogManager.getLogger(BlockLuminanceFixer.class);
 
 	@Override
 	public GeneratableElement convert(Workspace workspace, GeneratableElement input, JsonElement jsonElementInput) {
-		Plant plant = (Plant) input;
+		Block block = (Block) input;
 		try {
 			if (jsonElementInput.getAsJsonObject().get("definition").getAsJsonObject().get("luminance") != null) {
 				double oldLuminance = jsonElementInput.getAsJsonObject().get("definition").getAsJsonObject()
 						.get("luminance").getAsDouble();
-				plant.luminance = (int) Math.floor(oldLuminance * 15);
+				block.luminance = (int) Math.floor(oldLuminance * 15);
 			}
 		} catch (Exception e) {
-			LOG.warn("Could not update luminance field of: " + plant.getModElement().getName());
+			LOG.warn("Could not update luminance field of: " + block.getModElement().getName());
 		}
-		return plant;
+		return block;
 	}
 
 	@Override public int getVersionConvertingTo() {
